shop_backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
